refactor: separa os métodos de autenticação para um AuthService
